requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
pyserial==3.5
requests==2.31.0
PyJWT==2.8.0
python-dotenv==1.0.0
psutil==5.9.5
asyncio
threading